requirements.txt
imutils==0.5.4
nmap==0.0.1
numpy==1.21.4
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
playsound==1.3.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqtdarktheme==1.0.3
pyserial==3.5
python-dotenv==0.19.2
schedule==1.1.0
scipy==1.7.3
